Added detection for generic block destruction to reduce number of functions executed every tick
If a machine only has a genric destruction method (and is either based on a dropper, dispenser, hopper or furnace)  their (generic_)aftermath fucntion is only called when tehy are being destroyed, not every tick. A generic_destruction function does not exist. The generic destruction conditions are now evaluated inside the genericblocktick function. the new specialblockdestruction function (formerly just referred to as blockdestruction) is still executed every tick.